Tests written for APIs
