+++ .pytest_cache/v/cache/lastfailed
@@ -1,0 +1,11 @@
+{
+  "test_book_a_meal.py::BookAMealTestCase::test_delete_meal": true,
+  "test_book_a_meal.py::BookAMealTestCase::test_delete_order": true,
+  "test_book_a_meal.py::BookAMealTestCase::test_edit_meal": true,
+  "test_book_a_meal.py::BookAMealTestCase::test_edit_order": true,
+  "test_book_a_meal.py::BookAMealTestCase::test_get_meal_by_id": true,
+  "test_book_a_meal.py::BookAMealTestCase::test_get_order_by_id": true,
+  "test_book_a_meal.py::BookAMealTestCase::test_meal_create": true,
+  "test_book_a_meal.py::BookAMealTestCase::test_menu_setup": true,
+  "test_book_a_meal.py::BookAMealTestCase::test_order_create": true
+}
